statusbar: Fix an issue where the message would resize the window
When the message was larger than what could be allocated for it,
the window would be resized to fit it.  Now the message is truncated
if it is too long, and the filename is hidden when a message is shown.
The window may still be forcefully resized due to the progressbar,
but that will only happen if the window is extremly small.