Added support for compiling with MinGW
As of now gstreamer is not being used when compiling for Windows. When
testing, gstreamer would cause crashes when creating thumbnails for webm
files. This might be fixed in gstreamer's 1.6 release.

This commit also includes some miscellaneous changes to the interface -
most of them won't be noticable.  One change that should be noticable is
that when no local image list is loaded the thumbnail bar will automatically
be hidden.