booru/imagelist: Prevent UI lockup when loading a new page
Before when loading a new page while the thumbnails of the previous
were still being loaded, the UI would lock up due the the thumbnail
thread being joined prematurely.  Now the old thread is cancelled
and joined before the new thread is created.
The UI will still lock up while the last thumbnail from the previous
page loads, but this is better than waiting for every thumbnail
to load before the UI is responsive.