imagelist: Make archive extracting multithreaded
This changes the way archive image lists are loaded.  The Archive::Image
class is now in charge of extracting it's corresponding image file.
Thumbnails are now loaded in a smarter order, based on the current index.
Also fixed an issue where ahoviewer wouldn't link if lbzip or libunrar
was disabled.