booru/page: Store current tags when executing a search
This fixes some odd behavior when the user would search and then change
the tags.  Loading new pages of the original search would reflect the
new tags - without the user having to submit the new tags.
Also made it so when a user enters no tags the tags are set to *.