statusbar: Improve usage of set_progress and set_message
Made priorities a little easier to understand.  set_progress is now
essentially an overload of set_message that sets the progress widgets
value, and tries to set the current message if it's priority is higher.

Tooltips > Messages > Save progress > Download progress, progress bar
will be shown and updated even when a tooltip or message is shown.