imagebox: Refactor GStreamer initialization
Attempts to use platform specific video sinks first, then falls back to
glimagesink.  Added error messages when things don't work.

There is a bug in the most recent release of GStreamer with the
glimagesink on X11 that causes it to not resize the rendering rectangle.
This bug is fixed in master and will be in 1.18 (whenever it is released).
From what I observed though xvimagesink is still the least cpu intensive
sink available on X11.