imagebox: Play video files as they download
This commit includes unused code that can be used to swap from playbin3
to the normal playbin gstreamer element.  I have tried for many hours to
reduce the memory usage of the gstreamer pipeline and have since given up.
Playbin3 provides the most responsive experience from my testing, but
also uses the most memory.

In the future the playbin version may be a setting option, and maybe
even the ability to disable the main part of this commit - streaming videos.

I've also enabled the ability to play mp4 files, this requires
gst-plugins-libav for avdec_h264/h265.  mp4's are the most common video
format on danbooru and sakugabooru so it was overdue.

The videosink element can now be overridden with the setting VideoSink.

Close #97