booru: Add an info box widget, which displays additional post info
This displays the post date, source, rating, and score.  By default it
will only appear when the mouse is hovered around the tag view, but this
can be disabled to always show it.  Dates are using a library which has
been accepted into the C++20 standard to convert dates into local time.
If the timezone data is unable to be fetched (it uses curl) times will
be displayed in UTC.