mainwindow: Make default window size monitor aware
This ensures that the window is positioned centered on the monitor the
window will be shown on (usually?) and sized based on the size of that monitor.

Not exactly sure if this is correct, but it works as intended with my
window manager.  This assumes that the window will initially be
displayed on the monitor where the mouse pointer is.