Add an optional plugin system via libpeas
Right now there are 2 types of plugins.  One being SiteActivatble plugin
which can implement any missing booru (or non booru) sites for use
withing the booru browser.

The other (WindowAbstract) is a generic plugin with minimal functionality
right now, but will gain more in the future.  This plugin type has the
ability to open files inside ahoviewer, have a menu item, and have a
keybinding assigned to it by the user.

There are a lot of other misc. changes in this commit that I'm too lazy
to separate into individual commits.