Add the ability to delete local files
By default this has a keybinding is Shift+Delete.  Confirmation
will be asked for before deleting.  An attempt will be made to move
the file to trash/recycle bin if supported, and if that fails the file
will be deleted permanently.
The confirmation dialog can be turned off from the preferences
dialog and from the confirmation dialog itself.

Deleting files from archives are not supported.

Close #119.