mainwindow: Use RGBA visual when available
This enables the use of a RGBA value for the imagebox background color,
if there's a compositor running.  The GTK theme might get in the way of
the transparency, so a css style for the main windows background color
might need to be set.

Also removed a old gtkmm init call that was left over from gtk2?