imagebox: Prefer hardware accelerated video decoders
The only downside that I have noticed with this is that some invalid or
malformed files that play fine with the libgstlibav plugin do not work
with the nvcodec plugin.  I looked and I couldn't find a way to modify
the pipeline on the fly to try a different decoder when one fails.

Don't attempt to use xvimagesink or ximagesink when using a rgba visual
- they don't work with it.

This commit also changes the default X11/Wayland sink to a gtkglsink.