image: Fix video thumbnails
For the longest time this has been an issue, but would only cause
crashes on Windows so I thought it was just a gstreamer bug on Windows -
when it was actually the thumbnail data being freed because
Gdk::Pixbuf::create_from_data doesn't copy the data.

Also added the large thumbnails directory as a directory to get
thumbanils from, although normal will be searched first.